requirements.txt
requests>=2.28.0
psycopg2-binary>=2.9.0
pytest>=7.0.0
pytest-cov>=4.0.0